requirements.txt
appdirs==1.4.2
bootstrap4==0.1.0
Django==3.0.3
django-heroku==0.3.1
django-bootstrap4==1.1.1
djangorestframework==3.11.0
python-decouple==3.3
packaging==16.8
pyparsing==2.1.10
psycopg2==2.7.3.2
gunicorn==20.0.4
whitenoise==3.3.1
SymPy==1.5.1
dj-database-url==0.5.0